get_authorities_aggregated_entry: try to pre-resolve entries seeds based on the sole external ids
to reach getEntryResolutionScore with external ids known to Wikidata already identified
